--- examples/experimental/i3s-17-and-debug/webpack.config.js
@@ -1,29 +1,0 @@
-const CONFIG = {
-  mode: 'development',
-
-  entry: {
-    app: './app.js'
-  },
-
-  output: {
-    library: 'App'
-  },
-
-  module: {
-    rules: [
-      {
-        // Transpile ES6 to ES5 with babel
-        // Remove if your app does not use JSX or you don't need to support old browsers
-        test: /\.js$/,
-        loader: 'babel-loader',
-        exclude: [/node_modules/],
-        options: {
-          presets: ['@babel/preset-env', '@babel/preset-react']
-        }
-      }
-    ]
-  }
-};
-
-// This line enables bundling against src in this repo rather than installed module
-module.exports = (env) => (env ? require('../../webpack.config.local')(CONFIG)(env) : CONFIG);
