+++ examples/website/i3s/bounding-volume-layer.js
