+++ examples/website/i3s/color-map.js
