+++ examples/website/i3s/components/attributes-panel.js
