+++ examples/website/i3s/helpers/build-minimap-data.d.ts
