+++ examples/website/i3s/helpers/build-minimap-data.js
