+++ examples/website/i3s/package.json
@@ -3,26 +3,43 @@
   "scripts": {
     "start": "webpack-dev-server --progress --hot --open",
     "start-local": "webpack-dev-server --env.local --env.esnext --progress --hot --open",
+    "start-local-debug": "webpack-dev-server --config webpack-debug.config.js --env.local --env.esnext --progress --hot --open",
     "start-local-deck": "webpack-dev-server --env.local --env.esnext --env.deck --progress --hot --open",
+    "start-local-debug-deck": "webpack-dev-server --config webpack-debug.config.js --env.local --env.esnext --env.deck --progress --hot --open",
     "start-local-luma": "webpack-dev-server --env.local --env.esnext --env.luma --progress --hot --open",
     "start-local-math": "webpack-dev-server --env.local --env.esnext --env.math --progress --hot --open",
     "generate": "node ./generate-index/index.js"
   },
   "dependencies": {
-    "@deck.gl/core": "8.4.6",
-    "@deck.gl/geo-layers": "8.4.6",
-    "@deck.gl/layers": "8.4.6",
-    "@deck.gl/mesh-layers": "8.4.6",
-    "@deck.gl/react": "8.4.6",
-    "@loaders.gl/i3s": "^2.3.12",
-    "@luma.gl/constants": "8.4.3",
-    "@luma.gl/core": "8.4.3",
+    "@deck.gl/core": "8.5.0-beta.1",
+    "@deck.gl/extensions": "8.5.0-beta.1",
+    "@deck.gl/geo-layers": "8.5.0-beta.1",
+    "@deck.gl/layers": "8.5.0-beta.1",
+    "@deck.gl/mesh-layers": "8.5.0-beta.1",
+    "@deck.gl/react": "8.5.0-beta.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.34",
+    "@fortawesome/free-solid-svg-icons": "^5.15.2",
+    "@fortawesome/react-fontawesome": "^0.1.14",
+    "@loaders.gl/core": "3.0.0-beta.8",
+    "@loaders.gl/draco": "3.0.0-beta.8",
+    "@loaders.gl/gltf": "3.0.0-beta.8",
+    "@loaders.gl/i3s": "3.0.0-beta.8",
+    "@loaders.gl/images": "3.0.0-beta.8",
+    "@luma.gl/constants": "8.5.0-beta.1",
+    "@luma.gl/core": "8.5.0-beta.1",
+    "@luma.gl/engine": "8.5.0-beta.1",
+    "@luma.gl/shadertools": "8.5.0-beta.1",
+    "@luma.gl/webgl": "8.5.0-beta.1",
     "@math.gl/core": "3.5.0-alpha.1",
+    "@math.gl/culling": "3.5.0-alpha.1",
     "@math.gl/geospatial": "3.5.0-alpha.1",
     "@probe.gl/stats-widget": "^3.3.0",
+    "gl-matrix": "^3.3.0",
+    "math.gl": "^3.4.2",
     "probe.gl": "^3.3.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.0",
+    "react-color": "^2.19.3",
     "react-dom": "^16.8.0",
     "react-map-gl": "^5.0.0",
     "styled-components": "^4.2.0"
